+++ src/lib/index.ts
@@ -1,9 +1,9 @@
 // public API for `@feltcoop/felt`:
 
-export * as icons from './lib/icons.js';
+export * as icons from './icons.js';
 
 // types
-export * from './lib/util/types.js';
+export * from './util/types.js';
 
 // TODO consider what top-level public API makes sense -
 // for now we're preferring deep imports to specific modules -
@@ -26,5 +26,5 @@ export * from './lib/util/types.js';
 // 	configure_log_level,
 // 	print_log_label,
 // } from './util/log.js';
-// export {Timings, create_stopwatch} from './util/time.js';
-// export type {Stopwatch} from './util/time.js';
+// export {Timings, create_stopwatch} from './util/timings.js';
+// export type {Stopwatch} from './util/timings.js';