+++ changelog.md
@@ -1,5 +1,19 @@
 # changelog
 
+## 0.3.0
+
+- **break**: swap the names of `spawn` and `spawn_process`
+  ([#63](https://github.com/feltcoop/felt/pull/63))
+- **break**: rename `spawn_restartable_process` from `create_restartable_process`
+  and remove its `delay` arg
+  ([#63](https://github.com/feltcoop/felt/pull/63))
+- **break**: rename `util/timings.ts` from `util/time.ts`
+  ([#63](https://github.com/feltcoop/felt/pull/63))
+- **break**: rename `Timings.get_all` from `Timings.getAll`
+  ([#63](https://github.com/feltcoop/felt/pull/63))
+- upgrade Gro to 0.28.0
+  ([#63](https://github.com/feltcoop/felt/pull/63))
+
 ## 0.2.2
 
 - fix another issue with published types